moved FavMessages related codes to one folder and added FavMessagesView Header
